+++ components/TextFields/examples/experimental/supplemental/MDCContainerStyleFilled.m
@@ -264,7 +264,7 @@ static const CGFloat kFilledContainerStyleTopCornerRadius = (CGFloat)4.0;
 }
 
 - (CGFloat)contentAreaTopPaddingWithFloatingPlaceholderMaxY:(CGFloat)floatingPlaceholderMaxY {
-  return floatingPlaceholderMaxY + 6.5;
+  return floatingPlaceholderMaxY + (CGFloat)6.5;
 }
 
 - (CGFloat)normalContentAreaBottomPadding {
